Ипотека [#156315 #226964]

+++ src/test/java/ru/job4j/loop/MortgageTest.java
@@ -21,7 +21,7 @@ public class MortgageTest {
     public void when2Year() {
         int amount = 100;
         int salary = 120;
-        int percent = 50;
+        double percent = 50;
         int result = Mortgage.year(amount, salary, percent);
         int expected = 2;
         Assert.assertEquals(expected, result);
